+++ include/prismspf/solvers/concurrent_explicit_solver.h
@@ -22,10 +22,7 @@ public:
    * @brief Constructor.
    */
   explicit ConcurrentExplicitSolver(const SolverContext<dim, degree> &_solver_context,
-                                    Types::Index                      _solve_priority = 0)
-    : ConcurrentSolver<dim, degree, number>(_solver_context,
-                                            FieldSolveType::Explicit,
-                                            _solve_priority) {};
+                                    Types::Index _solve_priority = 0);
 
   /**
    * @brief Destructor.
@@ -66,60 +63,25 @@ public:
    * @brief Initialize the solver.
    */
   void
-  init() override
-  {
-    // Call the base class init
-    this->ConcurrentSolver<dim, degree, number>::init();
-
-    // Do nothing
-  };
+  init() override;
 
   /**
    * @brief Reinitialize the solver.
    */
   void
-  reinit() override
-  {
-    // Call the base class reinit
-    this->ConcurrentSolver<dim, degree, number>::reinit();
-
-    // Do nothing
-  };
+  reinit() override;
 
   /**
    * @brief Solve for a single update step.
    */
   void
-  solve() override
-  {
-    // Call the base class solve
-    this->ConcurrentSolver<dim, degree, number>::solve();
-
-    // If the solver is empty we can just return early.
-    if (this->solver_is_empty())
-      {
-        return;
-      }
-
-    // Otherwise, solve
-    this->solve_explicit_equations(
-      [this](
-        std::vector<typename SolverBase<dim, degree, number>::VectorType *>       &dst,
-        const std::vector<typename SolverBase<dim, degree, number>::VectorType *> &src)
-      {
-        this->get_system_matrix()->compute_explicit_update(dst, src);
-      });
-  };
+  solve() override;
 
   /**
    * @brief Print information about the solver to summary.log.
    */
   void
-  print() override
-  {
-    // Print the base class information
-    this->ConcurrentSolver<dim, degree, number>::print();
-  }
+  print() override;
 };
 
 PRISMS_PF_END_NAMESPACE
