+++ include/prismspf/solvers/sequential_linear_solver.h
@@ -20,10 +20,7 @@ public:
    * @brief Constructor.
    */
   explicit SequentialLinearSolver(const SolverContext<dim, degree> &_solver_context,
-                                  Types::Index                      _solve_priority = 0)
-    : SequentialSolver<dim, degree, number>(_solver_context,
-                                            FieldSolveType::NonexplicitLinear,
-                                            _solve_priority) {};
+                                  Types::Index                      _solve_priority = 0);
 
   /**
    * @brief Destructor.
@@ -64,77 +61,25 @@ public:
    * @brief Initialize the solver.
    */
   void
-  init() override
-  {
-    // Call the base class init
-    this->SequentialSolver<dim, degree, number>::init();
-
-    // If the solver is empty we can just return early.
-    if (this->solver_is_empty())
-      {
-        return;
-      }
-
-    // Init the linear solvers
-    for (const auto &[index, variable] : this->get_subset_attributes())
-      {
-        this->init_linear_solver(variable);
-      }
-  };
+  init() override;
 
   /**
    * @brief Reinitialize the solver.
    */
   void
-  reinit() override
-  {
-    // Call the base class reinit
-    this->SequentialSolver<dim, degree, number>::reinit();
-
-    // If the solver is empty we can just return early.
-    if (this->solver_is_empty())
-      {
-        return;
-      }
-
-    // Reinit the linear solvers
-    for (const auto &[index, variable] : this->get_subset_attributes())
-      {
-        this->reinit_linear_solver(variable);
-      }
-  };
+  reinit() override;
 
   /**
    * @brief Solve for a single update step.
    */
   void
-  solve() override
-  {
-    // Call the base class solve
-    this->SequentialSolver<dim, degree, number>::solve();
-
-    // If the solver is empty we can just return early.
-    if (this->solver_is_empty())
-      {
-        return;
-      }
-
-    // Solve each field
-    for (const auto &[index, variable] : this->get_subset_attributes())
-      {
-        this->solve_linear_solver(variable);
-      }
-  };
+  solve() override;
 
   /**
    * @brief Print information about the solver to summary.log.
    */
   void
-  print() override
-  {
-    // Print the base class information
-    this->SequentialSolver<dim, degree, number>::print();
-  }
+  print() override;
 };
 
 PRISMS_PF_END_NAMESPACE
