+++ src/solvers/concurrent_explicit_solver.inst.in
@@ -1,0 +1,4 @@
+for ( dimension : SPACE_DIMENSIONS; degree : ELEMENT_DEGREE; number : REAL_SCALARS)
+  {
+    template class ConcurrentExplicitSolver<dimension, degree, number>;
+  }
