+++ src/solvers/sequential_co_nonlinear_solver.inst.in
@@ -1,0 +1,4 @@
+for ( dimension : SPACE_DIMENSIONS; degree : ELEMENT_DEGREE; number : REAL_SCALARS)
+  {
+    template class SequentiaSequentialCoNonlinearSolverlAuxiliarySolver<dimension, degree, number>;
+  }
